requirements.txt
alembic==0.8.6
alpha-vantage==1.3.6
amqp==2.2.2
aniso8601==1.1.0
appnope==0.1.0
arrow==0.12.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
click==6.7
coverage==4.1
croniter==0.3.20
decorator==4.0.9
feedparser==5.2.1
Flask==0.10.1
Flask-HTTPAuth==3.1.2
flask-marshmallow==0.8.0
Flask-Migrate==1.8.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
future==0.16.0
gnureadline==6.3.8
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==4.1.0
lxml==4.1.1
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
nose==1.3.7
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.0
pexpect==4.3.0
pickleshare==0.7.4
psycopg2==2.7.3.2
ptyprocess==0.5.2
python-dateutil==2.5.3
python-editor==1.0.1
python-twitter==3.3
pytz==2016.4
redis==2.10.6
requests==2.10.0
requests-oauthlib==0.8.0
rq==0.9.2
rq-dashboard==0.3.10
rq-scheduler==0.7.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.10.0
sqlakeyset==0.1.1485813522
SQLAlchemy==1.0.13
traitlets==4.3.2
vine==1.1.4
Werkzeug==0.11.10
